+++ internal/converter/internal/otelcolconvert/converter_vcenterreceiver.go
@@ -92,6 +92,10 @@ func toVcenterMetricsConfig(cfg map[string]any) vcenter.MetricsConfig {
 		VcenterClusterMemoryLimit:          toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.memory.limit"])),
 		VcenterClusterVMCount:              toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.vm.count"])),
 		VcenterClusterVMTemplateCount:      toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.vm_template.count"])),
+		VcenterClusterVsanCongestions:      toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.vsan.congestions"])),
+		VcenterClusterVsanLatencyAvg:       toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.vsan.latency.avg"])),
+		VcenterClusterVsanOperations:       toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.vsan.operations"])),
+		VcenterClusterVsanThroughput:       toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.cluster.vsan.throughput"])),
 		VcenterDatacenterClusterCount:      toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.datacenter.cluster.count"])),
 		VcenterDatacenterCPULimit:          toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.datacenter.cpu.limit"])),
 		VcenterDatacenterDatastoreCount:    toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.datacenter.datastore.count"])),
@@ -115,6 +119,11 @@ func toVcenterMetricsConfig(cfg map[string]any) vcenter.MetricsConfig {
 		VcenterHostNetworkPacketDropRate:   toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.network.packet.drop.rate"])),
 		VcenterHostNetworkThroughput:       toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.network.throughput"])),
 		VcenterHostNetworkUsage:            toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.network.usage"])),
+		VcenterHostVsanCacheHitRate:        toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.vsan.cache.hit_rate"])),
+		VcenterHostVsanCongestions:         toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.vsan.congestions"])),
+		VcenterHostVsanLatencyAvg:          toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.vsan.latency.avg"])),
+		VcenterHostVsanOperations:          toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.vsan.operations"])),
+		VcenterHostVsanThroughput:          toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.host.vsan.throughput"])),
 		VcenterResourcePoolCPUShares:       toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.resource_pool.cpu.shares"])),
 		VcenterResourcePoolCPUUsage:        toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.resource_pool.cpu.usage"])),
 		VcenterResourcePoolMemoryBallooned: toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.resource_pool.memory.ballooned"])),
@@ -139,6 +148,9 @@ func toVcenterMetricsConfig(cfg map[string]any) vcenter.MetricsConfig {
 		VcenterVMNetworkPacketDropRate:     toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.vm.network.packet.drop.rate"])),
 		VcenterVMNetworkThroughput:         toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.vm.network.throughput"])),
 		VcenterVMNetworkUsage:              toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.vm.network.usage"])),
+		VcenterVMVsanLatencyAvg:            toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.vm.vsan.latency.avg"])),
+		VcenterVMVsanOperations:            toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.vm.vsan.operations"])),
+		VcenterVMVsanThroughput:            toVcenterMetricConfig(encodeMapstruct(cfg["vcenter.vm.vsan.throughput"])),
 	}
 }